+++ src/om_inputs/core.cljs
@@ -56,7 +56,7 @@
     (get-in opts [:type] (sch-type (:k-sch opts)))))
 
 (defn enum-label
-  "Display the label of a select or radio group entry"
+  "Display the i18n label of a select or radio group entry, fall backs to the code."
   [i18n code]
   (get-in i18n [:data code :label] (if (keyword? code) (full-name code) code)))
 
@@ -68,6 +68,8 @@
                 (dom/option #js {:value code}
                             (enum-label i18n code)))
               (:vs k-sch))))
+
+
 (defn radio-group
   [style {{:keys [attrs k i18n k-sch]} :opts chan :chan}]
   (apply dom/div #js {:className "input-group"}
@@ -92,39 +94,51 @@
   (radio-group "radio-inline" m))
 
 
+(defn make-segmented
+  "HOF, generates a function,  that closes over the value, for segemented control"
+  [type k value i18n chan]
+  (fn [code]
+    (dom/button #js {:type      type
+                     :active    (= code value)
+                     :className (styles "btn" (if (= code value) "btn-primary" "btn-default"))
+                     :key       (str (full-name k) "/" code)
+                     :id        (str (full-name k) "/" code)
+                     :value     code
+                     :onClick   #(put! chan [k code])}
+                (enum-label i18n code))))
+
 (defmethod magic-input "btn-group"
   [{{:keys [attrs k i18n k-sch]} :opts chan :chan}]
-  (apply dom/div #js {:className "btn-group"
-                      :id        (full-name k)}
-         (map (fn [code]
-                (dom/button (clj->js (merge attrs {:type      "button"
-                                                   :active    (= code (:value attrs))
-                                                   :className (styles "btn" (if (= code (:value attrs)) "btn-primary" "btn-default"))
-                                                   :key       (str (full-name k) "/" code)
-                                                   :id        (str (full-name k) "/" code)
-                                                   :value     code
-                                                   :onClick   #(put! chan [k code])}))
-                            (enum-label i18n code)))
-              (:vs k-sch))))
+  (apply dom/div (clj->js (merge attrs {:className "btn-group"}))
+         (map
+           (make-segmented "button" k (:value attrs) i18n chan)
+           (:vs k-sch))))
 
 (defmethod magic-input "range-btn-group"
-  [{{:keys [attrs k i18n k-sch]} :opts chan :chan}]
-  (let [min (int (:min attrs))
-        max (inc (int (:max attrs)))]
-   (apply dom/div #js {:className "btn-group"
-                       :id (full-name k)}
-          (map (fn [code]
-                 (dom/button (clj->js (merge attrs {:type      "button"
-                                                    :active    (= code (:value attrs))
-                                                    :className (styles "btn" (if (= code (:value attrs)) "btn-primary" "btn-default"))
-                                                    :key       (str (full-name k) "/" code)
-                                                    :id        (str (full-name k) "/" code)
-                                                    ;  :name (full-name k)
-                                                    :value     code
-                                                    :onClick   #(put! chan [k code])}))
-                             code))
-               (range min max)))))
+  [{{:keys [attrs i18n k]} :opts chan :chan}]
+  (let [{:keys [min max step value] :or {step 1}} attrs]
+    (apply dom/div (clj->js (merge attrs {:className "btn-group"}))
+           (map
+             (make-segmented "button" k value i18n chan)
+             (range (int min) (inc (int max)) step)))))
 
+(defmethod magic-input "stepper"
+  [{{:keys [attrs k]} :opts chan :chan}]
+  (let [{:keys [min max step value]} attrs
+        plus (if step (partial + (long step)) inc)
+        minus (if step (partial + (- (long step))) dec)]
+    (prn (minus value))
+   (dom/div (clj->js (merge attrs {:className "btn-group"}))
+            (dom/button #js {:type      "button"
+                             :className "btn btn-default"
+                             :onClick #(when (or (nil? min)
+                                                 (and min (<= (int min) (minus value))))
+                                        (put! chan [k (minus value)]))} "-")
+            (dom/button #js {:type      "button"
+                             :className "btn btn-default"
+                             :onClick #(when (or (nil? max)
+                                                 (and max (<= (plus value) (int max))))
+                                        (put! chan [k (plus value)]))} "+"))))
 
 
 (defmethod magic-input s/Inst
@@ -316,8 +330,8 @@
      [_ state]
      (let [chan-name (keyword (str (name k) "-chan"))
            chan (om/get-state owner chan-name)
-           state (get-in state [:action-state k])
-           btn-style (name state)]
+           button-state (get-in state [:action-state k])
+           btn-style (when button-state (name button-state))]
        (dom/div nil
                 (dom/button #js {:type  "button"
                                  :id (str (full-name comp-name) "-" (name k))