+++ CHANGELOG.md
@@ -1,9 +1,13 @@
+## 0.3.7-SNAPSHOT
+
+* ADD new input diplay : stepper -/+ for s/Int
+* ADD Possibility to put an HTML description to a field with sablono.
+
 ## 0.3.6-SNAPSHOT
 
 * ADD Layout indirection for input
 * ADD New input display segmented control with : btn-group and radio-btn-group
 * TEC Rewrite the way inputs are build. Each input is now given only its data.
-* ADD Possibility to put an HTML description to a field with sablono.
 
 
 ## 0.3.5-SNAPSHOT